% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grove.R
\docType{package}
\name{grove}
\alias{grove}
\alias{grove-package}
\title{grove: A package for functional denoising and functional ANOVA}
\description{
The grove package implements a wavelet-domain Bayesian hierarchical model for
functional analysis of variance.
}

