% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invDWT.R
\name{invDWT}
\alias{invDWT}
\title{Inverse discrete wavelet transform}
\usage{
invDWT(grove_obj, x, include_C = TRUE, sample_C = FALSE)
}
\arguments{
\item{grove_obj}{An object of class \code{grove}.}

\item{x}{A vector with the value of the predictor.}

\item{include_C}{If \code{TRUE}, C is used to compute to reconstruct the function.}

\item{sample_C}{If \code{TRUE}, draws from C are used to recontruct the function.}
}
\value{
A matrix each raw represents a draw from the reconstructed signal.
}
\description{
This function performs the inverse discrete wavelet transform.
}
\examples{
a <- 'TO DO'
}

