% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fANOVA.R
\name{fANOVA}
\alias{fANOVA}
\title{Bayesian functional ANOVA}
\usage{
fANOVA(W, formula, X, alpha = 0.5, beta = 1, nu = 5, n_samp = 500,
  transition_mode = "Markov", method = "Nelder-Mead")
}
\arguments{
\item{W}{An object of class \code{DWT}.}

\item{alpha}{Hyperparameter.}

\item{beta}{Hyperparameter.}

\item{nu}{Hyperparameter.}

\item{n_samp}{Number of posterior draws.}

\item{transition_mode}{Type of transition. The two options are \code{Markov} or \code{Independent}.}

\item{method}{Method used for find maxmimum of marginal likelihood.}

\item{eta}{Hyperparameter.}

\item{gamma}{Hyperparameter.}

\item{is.alpha.fixed}{If \code{TRUE}, alpha is fixed. If \code{FALSE} alpha is defined using Empirical Bayes
starting from value at \code{alpha}.}

\item{is.eta.fixed}{If \code{TRUE}, eta is fixed. If \code{FALSE} eta is defined using Empirical Bayes.}

\item{is.gamma.fixed}{If \code{TRUE}, gamma is fixed. If \code{FALSE} gamma is defined using Empirical Bayes.}
}
\value{
An object of class \code{grove}.
}
\description{
This function carries out Bayesian functional ANOVA using the Normal Inverse Gamma Markov Grove method of
Ma and Soriano (2016).
}
\examples{
a <- 'TO DO'
}

