% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFun.R
\name{plotFun}
\alias{plotFun}
\title{Function to plot the denoised signal}
\usage{
plotFun(data, p = c(0.025, 0.5, 0.975), main = "", col = "blue",
  type = "l", ylab = "", xlab = "", ylim = NULL, ...)
}
\arguments{
\item{data}{}

\item{p}{}

\item{main}{}

\item{col}{}

\item{type}{}

\item{ylab}{}

\item{xlab}{}

\item{ylim}{}
}
\value{
A plot.
}
\description{
This function plot the denoised signal
}
\examples{
a <- 'TO DO'
}

