% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotStates.R
\name{plotStates}
\alias{plotStates}
\title{Function to plot the hidden states}
\usage{
plotStates(grove_obj, block = "Intercept", legend = FALSE, main = NULL,
  prior = FALSE, ...)
}
\arguments{
\item{grove_obj}{}

\item{block}{}

\item{legend}{}

\item{main}{}

\item{prior}{}
}
\value{
A plot.
}
\description{
This function plot on a tree the state of the latent variables.
}
\examples{
a <- 'TO DO'
}

